Refactor tacacs_creds.yaml
