+++ tests/tacacs/test_accounting.py
@@ -8,10 +8,10 @@ import pytest
 
 from .test_authorization import ssh_connect_remote_retry, ssh_run_command, \
         remove_all_tacacs_server
-from tests.common.helpers.tacacs_helper import stop_tacacs_server, start_tacacs_server, \
+from tests.common.helpers.tacacs.tacacs_helper import stop_tacacs_server, start_tacacs_server, \
     per_command_accounting_skip_versions
 from .utils import check_server_received, change_and_wait_aaa_config_update, get_auditd_config_reload_timestamp, \
-        ensure_tacacs_server_running_after_ut  # noqa: F401
+    ensure_tacacs_server_running_after_ut    # noqa: F401
 from tests.common.errors import RunAnsibleModuleFail
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import skip_release
